requirements.txt
Flask==2.3.3
firebase-admin==6.2.0
requests==2.31.0
python-dotenv==1.0.0
Flask-Login==0.6.3

